feat(dictionary): move dictionary logic to server-side dir to implicate usage
