+++ src/app/[lang]/account/page.tsx
@@ -1,4 +1,4 @@
-import { getDictionary } from '@/get-dictionary';
+import { getDictionary } from "@/server/get-dictionary";
 import AccountClient from './account-client';
 import { type Locale } from '@/i18n-config';
 import { type Metadata } from 'next';
