+++ src/app/[lang]/report/[type]/page.tsx
@@ -2,7 +2,7 @@ import { HydrateClient } from "@/trpc/server";
 import { type Locale } from "@/i18n-config";
 import { notFound, redirect } from "next/navigation";
 import { ReportingForm } from "@/components/reporting/reporting-form";
-import { getDictionary } from "@/get-dictionary";
+import { getDictionary } from "@/server/get-dictionary";
 import { auth } from "@/server/auth";
 import { logger } from "@/lib/logger";
