+++ src/app/[lang]/terms/page.tsx
@@ -1,4 +1,4 @@
-import { getDictionary } from "@/get-dictionary";
+import { getDictionary } from "@/server/get-dictionary";
 import { type Locale } from "@/i18n-config";
 import { Footer } from "@/components/footer";
