+++ src/server/api/routers/report.ts
@@ -5,7 +5,7 @@ import { pictures } from "@/server/db/schema/pictures"
 import { protocolls } from "@/server/db/schema/protocoll"
 import { inArray } from "drizzle-orm"
 import { eq, and, min } from "drizzle-orm"
-import { getDictionary } from "@/get-dictionary"
+import { getDictionary } from "@/server/get-dictionary";
 import type { Locale } from "@/i18n-config"
 import { sendCreationNotification } from "@/lib/sendCreationNotification"
 import { env } from "@/env"
