+++ src/server/get-dictionary.ts
@@ -1,16 +1,16 @@
 import "server-only";
-import type { Locale } from "./i18n-config";
+import type { Locale } from "../i18n-config";
 
 type AllowedLanguage = "de" | "en";
 
 // Define the structure of your dictionary
-import { type Dictionary } from "./dictionaries/dictionary";
+import { type Dictionary } from "../dictionaries/dictionary";
 
 // We enumerate all dictionaries here for better linting and typescript support
 // We also get the default import for cleaner types
 const dictionaries: Record<AllowedLanguage, () => Promise<Dictionary>> = {
-    de: () => import("./dictionaries/de.json").then((module): Dictionary => module.default),
-    en: () => import("./dictionaries/en.json").then((module): Dictionary => module.default),
+    de: () => import("../dictionaries/de.json").then((module): Dictionary => module.default),
+    en: () => import("../dictionaries/en.json").then((module): Dictionary => module.default),
 
 };
 
@@ -25,7 +25,7 @@ export const getDictionary = async (locale: Locale): Promise<Dictionary> => {
 
     // Return cached version if available
     if (dictionaryCache[lang]) {
-        return dictionaryCache[lang]!;
+        return dictionaryCache[lang];
     }
 
     // Load and cache if not available
